+++ NEWS.md
@@ -1,5 +1,10 @@
 # anthro (development version)
 
+## Bugfix
+
+* Z-scores are now only computed for `age < 60 month` instead of `age <= 60`.
+* Z-scores for wfl were previously also computed when `age >= 60` months and `<= 1856 days`.
+
 # anthro 0.9.2
 
 * Age in days is now rounded half to even (e.g. 730.5 days = 731) before joining